Fix widgetgallery surfacegraph crash
When selecting a point on the surface graph
and then changing to slice selection,
the slice view is not created. Then selecting
a slice point tries to update the slice view
which is not created. The solution: when updating
the slice view, create slice view if it's not
created already

Fixes: QTBUG-116398
Change-Id: I63f0937f52450a72ed66828e056582c89c90e49a
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
(cherry picked from commit 7156846862adfc4afb956670cded2c790f6bfda9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>