Fix invisible custom labels
There was a rounding error (int/float), and scaling was
too small for them to show up. Changed the scaling logic
to be based on the height of the label instead of width,
as custom labels cannot be assumed to be of the same width.

Fixes: QTBUG-116525
Change-Id: Id10adb3a6ea5ee0df6bc79f84f03215e10fe20b0
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Dilek Akcay <dilek.akcay@qt.io>
Reviewed-by: Kwanghyo Park <kwanghyo.park@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
(cherry picked from commit 70d1a865bbd03e9ae53a01326751b89ef0760903)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>