Fix slice black triangle
The original fix for the oscilliscope was incorrect.
The reason for the crash (and other potential crashes)
was not clamping the ranges to 0 and 1 which allowed for column and row
limits below 0 and above the total count.

Fixes: QTBUG-116514
Change-Id: Ibb1ab3dbf33aab616293856e8389629a52321740
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
(cherry picked from commit 00a775f428d7865be754a29dfc63d1e8aaa2b3d6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>