Fix multiseries picking
Fixes the highlight series slice graph for the topography example as well
as multiseries picking for the surface test.
This commit also changes the coordinate system for surface graphs from
(i, j) to (j,i) to better reflect the rows and columns so that i = rows
= z-axis and j = columns = x-axis. This requires the topography example
to be slightly altered.

Fixes: QTBUG-116785
Change-Id: I0c90ef711a3b2e2bc725012feebc79b2bd3e4e51
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>