Remove unused d-pointer from scatter data item
It was deemed to be unnecessary since the class already has almost
everything it requires. However, the class might need to have one
expansion in the future. The expansion, if it's added, is going to be
scale. The scale can be expressed as a float. Hence, quintptr reserved
in the class.

Task-number: QTBUG-118169
Change-Id: Id7ad30271d2b0be2d83ce3a7dd9944b112719d7f
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>