Fix 'Zoom to selected bar' button
It was determined that the reason why graph didn't take the expected position when 'Zoom to selected bar' button was clicked due to the lack of needRender signal. These missing signals-slots existed in setActiveCamera function
in Q3DScene class before Q3DCamera was removed. The missing ones are now added in setCameraPreset function. Apart from this, a few minor changes were made in widgetgraphgallery's code to make the zoomed graph resemble the old one.

Fixes: QTBUG-118039
Change-Id: I8640544db2d03436bd12238de6f170ce69294a81
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>