Rename 3D graphs of QtGraphs module to QtGraphs3D
After renaming it matches "import QtQuick3D". This "3D" in the end
informs users that this module is using Qt Quick 3D.

This change is done in anticipation of the addition of 2D graphs
in the QtCharts re-implementation into QtGraphs module. 2D graphs
will then be imported into QML with "import QtGraphs". This matches
other 2D modules in QtQuick, which also do not state implicitly
being 2D.

Fixes: QTBUG-118235
Change-Id: Ib712776a04b1a61c8e266c7084868bfe43d9b9df
Reviewed-by: Kwanghyo Park <kwanghyo.park@qt.io>
Reviewed-by: Sakaria Pouke <sakaria.pouke@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
Reviewed-by: Dilek Akcay <dilek.akcay@qt.io>