Fix empty itemLabel on surfacegallery example
The reason for creating an empty itemLabel in spectogram example is that x and y values of selectedVertex.coord, row and column values basically, were set in reverse compared to the old implementation. Since they were set in reverse, if condition in setSelectedPoint sets pos as invalidSelectionPosition like x and y values exceed maxRow&maxCol. This has been fixed. It can also be seen from other examples that x and z values are written incorrectly in itemLabel (e.g: height map on surfacegraphgallery)

Fixes: QTBUG-116602
Change-Id: Iac66110a4a09efd2c611f9e36eceec3a6cdb7f32
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>