Fix angular grid for SurfaceSpectrogram
Angular grid happened to look correct for that one example. However,
when the line count was changed, the lines rotation was incorrect since
the rotation point was on the middle of the line. The algorithm expects
that the rotation point is on the left. That can be achieved by moving
the line model by the bounds.minimum.x along the X-axis in the gridline
component.

Fixes: QTBUG-119617
Change-Id: Iea45be3bd45af03c0e036a1b0ed4c43d509dd7d5
Reviewed-by: Dilek Akcay <dilek.akcay@qt.io>
Reviewed-by: Sakaria Pouke <sakaria.pouke@qt.io>
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>