Change colors when platform color scheme changes
Add a connection to listen to platform color changes. This is done by
adding an internal color scheme enum that contains Automatic value
in addition to Unknown, Light and Dark. The user can opt in and out
of the automatic behavior by setting the color scheme with the new
enum.

Fixes: QTBUG-125782
Pick-to: 6.8
Change-Id: I21a82269d498b1fc17c675bcedae7fd71bc36e1c
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>