Fix animation issue in directional test
Removed the condition of checking the selected item in triggerRotation() function as it wasn't working in the datavis3D, we don't need to set the selected item and it's rotation separately and more importantly, this only calculated the rotation of selected item, not all of them.

For the second issue of this bug we need to update items positions when both the m_isDataDirty and m_isSeriesVisualsDirty flags are set to true as positions of them also related to the visual.

Fixes: QTBUG-127769
Pick-to: 6.8.0
Change-Id: I55322975958574b3868e771d4a971e2caf8c8813
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
(cherry picked from commit 25965e8474ea653933aa304aa79f8c2ca879979a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>