Add clipPlotArea property for GraphsView in 2D
For too long, developers were bound by an unyielding rule:
Every graph item that dared to venture beyond the plot area was
mercilessly clipped, with no means to defy this constraint.

But now, everything changes.

With the arrival of this patch, a new power emerges — a property named
clipPlotArea. This simple yet formidable addition grants developers the
ong-sought ability to choose:
Should the outliers be silenced... or set free?

The control is finally in your hands.

Task-number: QTBUG-135887
Change-Id: Id1d4b5b3fa4405b23bfe606ef83d2712dcd15d27
Reviewed-by: Alexei Cazacov <alexei.cazacov@qt.io>
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>