GraphsWidgets: Do not introduce new QSharedPointers
Based on API review: Do not introduce new APIs with QSharedPointer

Rationale: QSharedPointer's implementation is poor and requires
twice the number of atomic operations on copy than std::shared_ptr.
We want to remove the class in Qt 7.

Solution: Instead of returning shared pointer to a QQuickItemGrabResult
that is received from the QQuick3DViewport, extract the image from
the grab result and return a pointer to that instead.

Pick-to: 6.10
Task-number: QTBUG-137478
Change-Id: I54cc9aece72cf8cf5ec4abaddc1da189d0c31b2b
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Kwanghyo Park <kwanghyo.park@qt.io>