Fix: crash when closing Graphs2D
When series is created outside of graphsview and then added into
a graphsview, the app crashes while closing. This seems to be because
it ends up calling pure virtual type function from the QAbstractSeries
in the getSeriesRendererIndex. This patch adds type variable to
QAbstractSeriesPrivate. The variable is set by QAbstractSeriesPrivate
constructor. The getSeriesRendererIndex function checks the value of
that variable instead of calling the virtual type().

Pick-to: 6.8 6.9
Fixes: QTBUG-138506
Change-Id: I7aca4970c38cbf83413818b0a77e7cf5d76195b6
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
(cherry picked from commit 630b38b3eed20c654cc13f5aedae4f704adfd963)
Reviewed-by: Sami Varanka <sami.varanka@qt.io>