Make 'qtquick3d' optional dependency for the module
QtQuick3D was listed as a required module in dependencies. However,
you should be able to build QtGraphs without 3D features. In that case,
it is not necessary to have Quick3D. Therefore, the Quick3D required
field should be false in the dependencies. In addition, the baseline
tests must be turned off if 3D features are unavailable, as they link
to Quick3D.

Pick-to: 6.9 6.8
Change-Id: I6c45fde93d75d54b4adf89471441d989452b0245
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
(cherry picked from commit 6300f5af2f58669306531a41e07c80560bbb6abb)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>