QAbstract3DSeries: don't fully qualify enum in member functions
The enum from the same class can be used in member functions without
fully qualifying the type with the class name. This makes a difference
for string-based signal/slot connections, where the declaration without
the full scope makes connections also work if the full scope is not
provided in the SIGNAL and SLOT macros.

Addresses header review comment and amends
52a361988d84ed893ac4cea36fe11e49d2c58fa5.

As a drive-by, fix the capitalization of the setter parameter, and the
documentation of the property.

Pick-to: 6.10
Change-Id: I1d9ca21dc8fd173852ca89ec3060dbb1da50c046
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Sami Varanka <sami.varanka@qt.io>