API review: don't return pointer to QImage from renderSliceToImage
In fact, don't return the QImage at all. It's an asynchronous function,
and the rendering will take an unknown time to complete. Callers must
connect to the sliceImageChanged() anyway to get the result, so just
let them do that, instead of clumsily polling with a timer.

Update the documentation and manual tests as much as possible for
now. There are baseline tests, but no auto-tests, so hard to say if this
still works correctly. It's likely that this needs a bit more work as a
follow up, but at least we don't release any dangerous API.

Pick-to: 6.10
Change-Id: I19d616c76746a826c4a46ed1df0aab4dc988556c
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>