Add custom series
Add a way for the user to create custom graph types. This implements
CustomSeries and CustomRenderer. The CustomSeries is a collection
of data points that the user can define and has a QML component delegate
that is used by CustomRenderer to create QML items that inherit
properties from the custom data.

Change-Id: I5925fc6404d82a328c106cab784bd2176e4167b9
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Sami Varanka <sami.varanka@qt.io>