Fix: Graphs3D series item label visibility
Bars3D and Scatter3D didn't hide the item label according to series'
itemLabelVisible. This was also in the Surface3D but it was fixed
already. However, the fix was not complete since it did not work if the
selection was active.

Fixes: QTBUG-141370
Pick-to: 6.8
Change-Id: I65acfc33e0d41cd4322efe521b7d7ba509ca159a
Reviewed-by: Niko Korkala <niko.korkala@qt.io>
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
(cherry picked from commit 8d8f2ffc91a672eb378f53691e77b7c758a73377)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>